Streamline scripts for fetching locale data
