+++ package.json
@@ -12,7 +12,7 @@
   "author": "sebworks",
   "private": true,
   "scripts": {
-    "dev": "npm run i18n:get-valid-locales && cross-env ENABLE_AUDIO=true nuxt",
+    "dev": "npm run i18n:create-locales-list && cross-env ENABLE_AUDIO=true nuxt",
     "build": "npm run i18n:get-translations && nuxt build",
     "generate": "nuxt generate",
     "start": "nuxt start",
@@ -32,12 +32,10 @@
     "format:test": "prettier --check .",
     "vue-i18n-extract": "vue-i18n-extract report -v './src/**/*.?(js|vue)' -l './src/locales/*.?(json)'",
     "prepare": "husky install",
-    "i18n:get-i18n-locales": "node src/locales/scripts/get-validated-locales",
-    "i18n:get-translations": "node src/locales/scripts/get-translations && npm run i18n:get-valid-locales",
+    "i18n:update-locales": "node src/locales/scripts/get-validated-locales",
+    "i18n:get-translations": "node src/locales/scripts/get-translations && npm run i18n:update-locales",
     "i18n:generate-pot": "node src/locales/scripts/json-to-pot",
-    "i18n:create-locales-list": "node src/locales/scripts/create-locale-list",
-    "i18n:update-locales": "node src/locales/scripts/update-locale-list",
-    "i18n:get-translated-status": "node src/locales/scripts/get-translations-status"
+    "i18n:create-locales-list": "node src/locales/scripts/create-wp-locale-list"
   },
   "dependencies": {
     "@nuxt/components": "^2.1.6",
