+++ src/locales/scripts/types.d.ts
@@ -1,3 +1,3 @@
 /**
- * @typedef {{code: string, wpLocale?: string, file?: string, iso?: string, name: string, dir?: string}} i18nLocaleProps
+ * @typedef {{ code: string, wpLocale?: string, file?: string, iso?: string, name: string, dir?: string, translated?: number }} i18nLocaleProps
  */
